APIS/Google_Ads_Api/requirements.txt
google-ads==21.2.0
google-auth 
google-auth-oauthlib 
google-auth-httplib2
